Added header and program files for the card struct. Added a player class. Error running code
